Cleaned up readme and docs. refactored _mock file into a module
