+++ docs/source/index.rst
@@ -1,7 +1,7 @@
 .. typemock documentation master file, created by
-   sphinx-quickstart on Tue Jun 11 20:45:13 2019.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
+sphinx-quickstart on Tue Jun 11 20:45:13 2019.
+You can adapt this file completely to your liking, but it should at least
+contain the root `toctree` directive.
 
 typemock: type safe mocking
 ===========================
@@ -40,7 +40,6 @@ Used in conjunction with mypy, this should result in much more high fidelity ind
    type_safety
    mocking_objects
    verifying
-   to_do
