--- docs/source/to_do.rst
@@ -1,16 +1,0 @@
-To Do and Issues
-================
-
-To Do:
-
-- Nested complex argument type safety. Recursively check attributes, not methods?
-- Handle mocking of functions (not methods)
-- Handle mocking of args, kwargs method signatures.
-- More control of type safety strictness through config. Not just an enum.
-- Helpers for mockng of static members? class methods, module variables etc...
-
-
-Issues:
-
-- Nothing here yet, but I am sure there will be soon.
-
