+++ tests/test_mock_class.py
@@ -1,6 +1,11 @@
 from unittest import TestCase
 
 from typemock import tmock
+from typemock.api import MockingError
+
+
+def a_static_function(func_arg: str) -> int:
+    pass
 
 
 class MyThingEmptyInit:
@@ -29,3 +34,7 @@ class TestMockClass(TestCase):
 
     def test_mock__class__init_with_args_and_complex_logic__no_errors(self):
         tmock(MyThingInitWithArgsAndLogic)
+
+    def test_mock_function__expect_error(self):
+        with self.assertRaises(MockingError):
+            tmock(a_static_function)
